+++ src/components/proposal/table/AddColumnModal.tsx
@@ -1,3 +1,0 @@
-import { AddColumnModal as SharedAddColumnModal } from '../../shared/modals/AddColumnModal';
-
-export const AddColumnModal = SharedAddColumnModal;