+++ src/components/proposal/table/AddGroupModal.tsx
@@ -1,3 +1,0 @@
-import { AddGroupModal as SharedAddGroupModal } from '../../shared/modals/AddGroupModal';
-
-export const AddGroupModal = SharedAddGroupModal;