+++ src/components/proposal/table/ChangeConfirmModal.tsx
@@ -1,7 +1,0 @@
-import {
-  ChangeConfirmModal as SharedChangeConfirmModal,
-  type ChangeConfirmResult,
-} from '../../shared/modals/ChangeConfirmModal';
-
-export const ChangeConfirmModal = SharedChangeConfirmModal;
-export type { ChangeConfirmResult };